feat(Game): make #render show conditional columns
The render method for the board will only print
column numbers of the board to the screen if a
move can be made in the corresponding column. If
no move is possible, the number of that column
won't be shown. This makes the state of the board
clearer to a human player.